backend/foodgram/requirements.txt
Django==4.2.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
djoser==2.2.0
Pillow==9.5.0
fpdf==1.7.2
django-filter==23.2
gunicorn==20.1.0